pyproject.toml
[project]
name = "deepseeker"
version = "0.1.0"
description = "Deep search & multi-LLM orchestration with OpenAI-style APIs"
authors = [{ name = "TabNahida", email = "tab@tabye.top" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
"httpx>=0.27",
"pydantic>=2.6",
"typer>=0.12",
"rich>=13.7",
"pyyaml>=6.0",
"anyio>=4.0",
"orjson>=3.10"
]


[project.scripts]
deepseeker = "deepseeker.cli:app"


[tool.pytest.ini_options]
testpaths = ["tests"]